+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 + Added "Free Camera" module
 + Fixed rotations overriding eachother with a priority system
 + Fixed "Anti Ghost Block" module from breaking blocks 
++ Fixed "Avoid" module not adding collision boxes to end portals
 ```
 
 ---